GameBot/mario_neuroevolution/requirements.txt
gym-super-mario-bros==7.4.0
nes-py==8.2.1
numpy==1.26.4
neat-python==0.92
matplotlib==3.8.2
opencv-python==4.9.0.80
pygame==2.5.2
setuptools==69.0.3
torch>=2.0.0
torchvision>=0.15.0